organise to standard file
